refactor: BigDecimal 금액 타입 변경에 맞춰 수익 흐름과 테스트 정리
- 컨트롤러의 배팅 금액 생성 로직을 BigDecimal 기반으로 수정
- GameResultManager의 딜러 수익 합산을 BigDecimal 연산으로 변경
- ParticipantRevenueDto의 수익 타입을 BigDecimal로 변경
- 배팅/수익 관련 테스트를 BigDecimal compareTo 방식으로 정리

+++ src/main/java/controller/BlackJackGameController.java
@@ -13,6 +13,7 @@ import domain.participant.Player;
 import domain.participant.Players;
 import dto.ParticipantCardsDto;
 import dto.ParticipantRevenueDto;
+import java.math.BigDecimal;
 import java.util.HashMap;
 import view.InputView;
 import view.OutputView;
@@ -50,7 +51,7 @@ public class BlackJackGameController {
         Map<Name, BettingAmount> bettingAmounts = new HashMap<>();
         players.forEach(player -> {
             int amount = InputView.askBettingAmount(player.getName().getName());
-            bettingAmounts.put(player.getName(), new BettingAmount(amount));
+            bettingAmounts.put(player.getName(), new BettingAmount(BigDecimal.valueOf(amount)));
         });
         return new BettingAmounts(bettingAmounts);
     }
