+++ src/main/java/dto/ParticipantRevenueDto.java
@@ -1,7 +1,9 @@
 package dto;
 
+import java.math.BigDecimal;
+
 public record ParticipantRevenueDto(
         String name,
-        int revenue
+        BigDecimal revenue
 ) {
 }
